@@ -1,6 +1,6 @@
 {
   "name": "react-native-snackbar-component",
-  "version": "1.0.8",
+  "version": "1.1.10",
   "description": "A snackbar component for Android and iOS",
   "main": "index.js",
   "scripts": {
@@ -9,7 +9,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/SiDevesh/React-Native-SnackBar.git"
+    "url": "git+https://github.com/sidevesh/react-native-snackbar-component.git"
   },
   "keywords": [
     "snackbar",
@@ -20,19 +20,19 @@
     "notifications",
     "react-native"
   ],
-  "author": "SiDevesh",
+  "author": "sidevesh",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/SiDevesh/React-Native-SnackBar/issues"
+    "url": "https://github.com/sidevesh/react-native-snackbar-component/issues"
   },
-  "homepage": "https://github.com/SiDevesh/React-Native-SnackBar#readme",
+  "homepage": "https://github.com/sidevesh/react-native-snackbar-component#readme",
   "devDependencies": {
     "babel-eslint": "^7.2.3",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^5.0.1",
-    "eslint-plugin-react": "^7.0.0"
+    "eslint": "^5.11.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.11.1"
   },
   "peerDependencies": {
     "react": ">=15.3.1",
